+++ app/src/main/java/com/nextcloud/talk/utils/message/MessageUtils.kt
@@ -9,13 +9,13 @@ package com.nextcloud.talk.utils.message
 import android.content.Context
 import android.content.Intent
 import android.graphics.Typeface
+import android.net.Uri
 import android.text.SpannableString
 import android.text.SpannableStringBuilder
 import android.text.Spanned
 import android.text.style.StyleSpan
 import android.util.Log
 import android.view.View
-import androidx.core.net.toUri
 import com.nextcloud.talk.R
 import com.nextcloud.talk.chat.data.model.ChatMessage
 import com.nextcloud.talk.ui.theme.ViewThemeUtils
@@ -28,6 +28,7 @@ import io.noties.markwon.ext.strikethrough.StrikethroughPlugin
 import io.noties.markwon.ext.tables.TablePlugin
 import io.noties.markwon.ext.tasklist.TaskListDrawable
 import io.noties.markwon.ext.tasklist.TaskListPlugin
+import com.nextcloud.talk.utils.MarkwonUtils
 
 class MessageUtils(val context: Context) {
     fun enrichChatReplyMessageText(
@@ -203,21 +204,7 @@ class MessageUtils(val context: Context) {
     }
 
     fun getRenderedMarkdownText(context: Context, markdown: String, textColor: Int): Spanned {
-        val drawable = TaskListDrawable(textColor, textColor, context.getColor(R.color.bg_default))
-        val markwon = Markwon.builder(context).usePlugin(object : AbstractMarkwonPlugin() {
-            override fun configureTheme(builder: MarkwonTheme.Builder) {
-                builder.isLinkUnderlined(true).headingBreakHeight(0)
-            }
-
-            override fun configureConfiguration(builder: MarkwonConfiguration.Builder) {
-                builder.linkResolver { view: View?, link: String? ->
-                    Log.i(TAG, "Link action not implemented $view / $link")
-                }
-            }
-        })
-            .usePlugin(TaskListPlugin.create(drawable))
-            .usePlugin(TablePlugin.create { _ -> })
-            .usePlugin(StrikethroughPlugin.create()).build()
+        val markwon = MarkwonUtils.build(context, textColor)
         return markwon.toMarkdown(markdown)
     }
